Modified but need work (see pundit-work.md)

+++ app/controllers/admin/permission_forms_controller.rb
@@ -82,12 +82,10 @@ class Admin::PermissionFormsController < ApplicationController
   public
 
   def index
-    # PUNDIT_REVIEW_AUTHORIZE
-    # PUNDIT_CHECK_AUTHORIZE
     authorize Admin::PermissionForm
     # PUNDIT_REVIEW_SCOPE
     # PUNDIT_CHECK_SCOPE (did not find instance)
-    @permission_forms = policy_scope(Admin::PermissionForm)
+    # @permission_forms = policy_scope(Admin::PermissionForm)
     form = TeacherSearchForm.new(params[:form])
     @teachers = form.search
   end
@@ -114,8 +112,6 @@ class Admin::PermissionFormsController < ApplicationController
   end
 
   def create
-    # PUNDIT_REVIEW_AUTHORIZE
-    # PUNDIT_CHECK_AUTHORIZE
     authorize Admin::PermissionForm
     form_data = params['portal_permission']
     if form_data && (!form_data['name'].blank?)
@@ -125,14 +121,8 @@ class Admin::PermissionFormsController < ApplicationController
   end
 
   def remove_form
-    # PUNDIT_REVIEW_AUTHORIZE
-    # PUNDIT_CHOOSE_AUTHORIZE
-    # no authorization needed ...
-    # authorize Admin::PermissionForm
-    # authorize @permission_form
-    # authorize Admin::PermissionForm, :new_or_create?
-    # authorize @permission_form, :update_edit_or_destroy?
     form = Portal::PermissionForm.find(params[:id])
+    authorize form, :destroy?
     form.destroy
     redirect_to action: 'index'
   end
