+++ app/controllers/portal/bookmarks_controller.rb
@@ -2,27 +2,19 @@ class Portal::BookmarksController < ApplicationController
   include Portal::BookmarksHelper
 
   def index
-    # PUNDIT_REVIEW_AUTHORIZE
-    # PUNDIT_CHECK_AUTHORIZE
     authorize Portal::Bookmark
     # This is needed so the side-menu selection works as expected.
     @portal_clazz = get_current_clazz
     @bookmarks = Portal::Bookmark.where(clazz_id: @portal_clazz)
     # PUNDIT_REVIEW_SCOPE
     # PUNDIT_CHECK_SCOPE (found instance)
-    @bookmarks = policy_scope(Portal::Bookmark)
+    # @bookmarks = policy_scope(Portal::Bookmark)
     # Save the left pane sub-menu item
     Portal::Teacher.save_left_pane_submenu_item(current_visitor, Portal::Teacher.LEFT_PANE_ITEM['LINKS'])
   end
 
   def add_padlet
-    # PUNDIT_REVIEW_AUTHORIZE
-    # PUNDIT_CHOOSE_AUTHORIZE
-    # no authorization needed ...
-    # authorize Portal::Bookmark
-    # authorize @bookmark
-    # authorize Portal::Bookmark, :new_or_create?
-    # authorize @bookmark, :update_edit_or_destroy?
+    authorize Portal::PadletBookmark, :new_or_create?
     mark = Portal::PadletBookmark.create_for_user(current_visitor, get_current_clazz)
     render :update do |page|
       page.insert_html :bottom,
@@ -33,13 +25,7 @@ class Portal::BookmarksController < ApplicationController
   end
 
   def add
-    # PUNDIT_REVIEW_AUTHORIZE
-    # PUNDIT_CHOOSE_AUTHORIZE
-    # no authorization needed ...
-    # authorize Portal::Bookmark
-    # authorize @bookmark
-    # authorize Portal::Bookmark, :new_or_create?
-    # authorize @bookmark, :update_edit_or_destroy?
+    authorize Portal::GenericBookmark, :new_or_create?
     props = params['portal_generic_bookmark'] || {
       name: 'My bookmark',
       url: 'http://concord.org'
@@ -57,27 +43,15 @@ class Portal::BookmarksController < ApplicationController
   end
 
   def visit
-    # PUNDIT_REVIEW_AUTHORIZE
-    # PUNDIT_CHOOSE_AUTHORIZE
-    # no authorization needed ...
-    # authorize Portal::Bookmark
-    # authorize @bookmark
-    # authorize Portal::Bookmark, :new_or_create?
-    # authorize @bookmark, :update_edit_or_destroy?
     mark = Portal::Bookmark.find(params['id'])
+    authorize mark, :show
     mark.record_visit(current_visitor)
     redirect_to mark.url
   end
 
   def delete
-    # PUNDIT_REVIEW_AUTHORIZE
-    # PUNDIT_CHOOSE_AUTHORIZE
-    # no authorization needed ...
-    # authorize Portal::Bookmark
-    # authorize @bookmark
-    # authorize Portal::Bookmark, :new_or_create?
-    # authorize @bookmark, :update_edit_or_destroy?
     mark = Portal::Bookmark.find(params['id'])
+    authorize mark
     dom_id = bookmark_dom_item(mark)
     if mark.changeable? current_visitor
       mark.destroy
@@ -112,10 +86,8 @@ class Portal::BookmarksController < ApplicationController
   end
 
   def edit
-    # PUNDIT_REVIEW_AUTHORIZE
-    # PUNDIT_CHECK_AUTHORIZE (did not find instance)
-    authorize @bookmark
     bookmark = Portal::Bookmark.find(params['id'])
+    authorize bookmark
     if bookmark && bookmark.changeable?(current_visitor)
       %w[name url is_visible].each do |param|
         unless params[param].blank?
